Refactor tool invocations test
