+++ tests/functions/test_openai.py
@@ -7,39 +7,14 @@ import pixeltable.exceptions as excs
 import pixeltable.functions as pxtf
 import pixeltable.type_system as ts
 
+from .tool_invocations import run_tool_invocations_test, server_state, stock_price, weather
 from ..conftest import DO_RERUN
 from ..utils import (
     SAMPLE_IMAGE_URL,
     skip_test_if_no_client,
     skip_test_if_not_installed,
-    stock_price,
     validate_update_status
 )
-
-
-@pxt.udf
-def weather(city: str) -> Optional[str]:
-    """
-    Get today's weather forecast for a given city.
-
-    Args:
-        city - The name of the city to look up.
-    """
-    if city == 'San Francisco':
-        return 'Cloudy with a chance of meatballs'
-    else:
-        return 'Unknown city'
-
-
-@pxt.udf
-def server_state() -> str:
-    """
-    Get the current server state.
-
-    Returns:
-        The current server state.
-    """
-    return 'Running (0x4171780)'
 
 
 @pytest.mark.remote_api
@@ -49,9 +24,9 @@ class TestOpenai:
     def test_audio(self, reset_db: None) -> None:
         skip_test_if_not_installed('openai')
         skip_test_if_no_client('openai')
-        t = pxt.create_table('test_tbl', {'input': pxt.String})
         from pixeltable.functions.openai import speech, transcriptions, translations
 
+        t = pxt.create_table('test_tbl', {'input': pxt.String})
         t.add_computed_column(speech=speech(t.input, model='tts-1', voice='onyx'))
         t.add_computed_column(
             speech_2=speech(
@@ -190,96 +165,17 @@ class TestOpenai:
         skip_test_if_no_client('openai')
         from pixeltable.functions.openai import chat_completions, invoke_tools
 
-        tools = pxt.tools(stock_price, weather)
-        tool_choice_opts: list[Optional[pxt.func.ToolChoice]] = [
-            None,
-            tools.choice(auto=True),
-            tools.choice(required=True),
-            tools.choice(tool='stock_price'),
-            tools.choice(tool=weather),
-            tools.choice(required=True, parallel_tool_calls=False),
-        ]
-
-        for tool_choice in tool_choice_opts:
-            pxt.drop_table('test_tbl', if_not_exists='ignore')
-            t = pxt.create_table('test_tbl', {'prompt': pxt.String})
+        def make_table(tools: pxt.func.Tools, tool_choice: pxt.func.ToolChoice) -> pxt.Table:
+            t = pxt.create_table('test_tbl', {'prompt': pxt.String}, if_exists='replace')
             messages = [{'role': 'user', 'content': t.prompt}]
             t.add_computed_column(
                 response=chat_completions(model='gpt-4o-mini', messages=messages, tools=tools, tool_choice=tool_choice)
             )
             t.add_computed_column(output=t.response.choices[0].message.content)
             t.add_computed_column(tool_calls=invoke_tools(tools, t.response))
+            return t
 
-            t.insert(prompt='What is the stock price of NVDA today?')
-            t.insert(prompt='What is the weather in San Francisco?')
-            t.insert(prompt='What is the stock price of NVDA today, and what is the weather in San Francisco?')
-            t.insert(prompt='How many grams of corn are in a bushel?')
-            t.insert(prompt='What is the stock price of NVDA today? Also, what is the stock price of UAL?')
-            res = t.select(t.response, t.tool_calls).head()
-            print(f'Responses with tool_choice equal to: {tool_choice}')
-            print(res[0]['response'])
-            print(res[1]['response'])
-            print(res[2]['response'])
-            print(res[3]['response'])
-            print(res[4]['response'])
-
-            # Request for stock price: works except when tool_choice is set explicitly to weather
-            print('Checking stock price inquiry')
-            if tool_choice is None or tool_choice.tool != 'weather':
-                assert res[0]['tool_calls'] == {'stock_price': [131.17], 'weather': None}
-            else:  # Explicitly set to weather; we may or may not get stock price also
-                assert res[0]['tool_calls'] in [
-                    {'stock_price': None, 'weather': ['Unknown city']},
-                    {'stock_price': [131.17], 'weather': ['Unknown city']},
-                ]
-
-            # Request for weather: works except when tool_choice is set explicitly to stock_price
-            print('Checking weather inquiry')
-            if tool_choice is None or tool_choice.tool != 'stock_price':
-                assert res[1]['tool_calls'] == {'stock_price': None, 'weather': ['Cloudy with a chance of meatballs']}
-            else:  # Explicitly set to stock_price; we may or may not get weather also
-                assert res[1]['tool_calls'] in [
-                    {'stock_price': [0.0], 'weather': None},
-                    {'stock_price': [0.0], 'weather': ['Cloudy with a chance of meatballs']},
-                ]
-
-            # Request for both stock price and weather
-            print('Checking double inquiry')
-            if tool_choice is None or (tool_choice.parallel_tool_calls and tool_choice.tool is None):
-                # Both tools invoked in parallel
-                assert res[2]['tool_calls'] == {
-                    'stock_price': [131.17],
-                    'weather': ['Cloudy with a chance of meatballs'],
-                }
-            elif tool_choice.tool == 'stock_price':
-                assert res[2]['tool_calls'] == {'stock_price': [131.17], 'weather': None}
-            elif tool_choice.tool == 'weather':
-                assert res[2]['tool_calls'] == {'stock_price': None, 'weather': ['Cloudy with a chance of meatballs']}
-            else:
-                # Only one tool invoked, but it's not specified which
-                assert not tool_choice.parallel_tool_calls
-                assert res[2]['tool_calls'] in [
-                    {'stock_price': [131.17], 'weather': None},
-                    {'stock_price': None, 'weather': ['Cloudy with a chance of meatballs']},
-                ]
-
-            print('Checking random question')
-            if tool_choice is None or tool_choice.auto:
-                assert res[3]['tool_calls'] == {'stock_price': None, 'weather': None}
-            elif tool_choice.tool == 'stock_price':
-                assert res[3]['tool_calls'] == {'stock_price': [0.0], 'weather': None}
-            elif tool_choice.tool == 'weather':
-                assert res[3]['tool_calls'] == {'stock_price': None, 'weather': ['Unknown city']}
-            else:
-                assert res[3]['tool_calls'] in [
-                    {'stock_price': [0.0], 'weather': None},
-                    {'stock_price': None, 'weather': ['Unknown city']},
-                ]
-
-            print('Checking multiple stock prices question')
-            if tool_choice is None or tool_choice.auto:
-                # If you specify an explicit tool, it seems to only call it once.
-                assert res[4]['tool_calls'] == {'stock_price': [131.17, 82.88], 'weather': None}
+        run_tool_invocations_test(make_table, test_tool_choice=True, test_individual_tool_choice=True)
 
     def test_custom_tool_invocations(self, reset_db: None) -> None:
         skip_test_if_not_installed('openai')
